+++ backend/canisters/local_user_index/impl/src/updates/register_user.rs
@@ -157,7 +157,7 @@ fn prepare(args: &Args, state: &mut RuntimeState) -> Result<PrepareOk, Response>
     };
 
     let canister_id = state.data.canister_pool.pop();
-    let canister_wasm = state.data.user_canister_wasm_for_new_canisters.clone();
+    let canister_wasm = state.data.user_canister_wasm_for_new_canisters.wasm.clone();
     let init_canister_args = InitUserCanisterArgs {
         owner: caller,
         group_index_canister_id: state.data.group_index_canister_id,